+++ src/netstandard/RaptorClassLibrary.netstandard/Analytics/TemperatureStatistics/GridFabric/TemperatureStatisticsResponse.cs
@@ -45,7 +45,7 @@ namespace VSS.TRex.Analytics.TemperatureStatistics.GridFabric
         AboveTargetPercent = ValueOverTargetPercent,
         TotalAreaCoveredSqMeters = SummaryProcessedArea,
 
-        ReturnCode = MissingTargetValue ? SummaryCellsScanned == 0 ? (short)MissingTargetDataResultType.NoResult : (short)MissingTargetDataResultType.PartialResult : (short)MissingTargetDataResultType.NoProblems,
+        ReturnCode = MissingTargetValue ? SummaryCellsScanned == 0 ? MissingTargetDataResultType.NoResult : MissingTargetDataResultType.PartialResult : MissingTargetDataResultType.NoProblems,
 
         ResultStatus = ResultStatus
       };
